requirements.txt
sqlmodel>=0.0.16
pymysql
cryptography
requests
dotenv
pydantic>=2.7.0
fastapi[standard]>=0.110.0
pwdlib[argon2]
pyjwt
pytest
